Split command into interface and implementation
