+++ src/main/java/net/dv8tion/jda/api/interactions/commands/Command.java
@@ -28,10 +28,7 @@ import net.dv8tion.jda.api.utils.TimeUtil;
 import net.dv8tion.jda.api.utils.data.DataArray;
 import net.dv8tion.jda.api.utils.data.DataObject;
 import net.dv8tion.jda.api.utils.data.DataType;
-import net.dv8tion.jda.internal.JDAImpl;
-import net.dv8tion.jda.internal.requests.RestActionImpl;
-import net.dv8tion.jda.internal.requests.Route;
-import net.dv8tion.jda.internal.requests.restaction.CommandEditActionImpl;
+import net.dv8tion.jda.internal.interactions.command.CommandImpl;
 import net.dv8tion.jda.internal.utils.Checks;
 
 import javax.annotation.CheckReturnValue;
@@ -39,8 +36,6 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.time.OffsetDateTime;
 import java.util.*;
-import java.util.function.Function;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 /**
@@ -50,50 +45,8 @@ import java.util.stream.Collectors;
  * @see Guild#retrieveCommandById(String)
  * @see Guild#retrieveCommands()
  */
-public class Command implements ISnowflake
+public interface Command extends ISnowflake
 {
-    private static final EnumSet<OptionType> OPTIONS = EnumSet.complementOf(EnumSet.of(OptionType.SUB_COMMAND, OptionType.SUB_COMMAND_GROUP));
-    private static final Predicate<DataObject> OPTION_TEST = it -> OPTIONS.contains(OptionType.fromKey(it.getInt("type")));
-    private static final Predicate<DataObject> SUBCOMMAND_TEST = it -> OptionType.fromKey(it.getInt("type")) == OptionType.SUB_COMMAND;
-    private static final Predicate<DataObject> GROUP_TEST = it -> OptionType.fromKey(it.getInt("type")) == OptionType.SUB_COMMAND_GROUP;
-
-    private final JDAImpl api;
-    private final Guild guild;
-    private final String name, description;
-    private final List<Option> options;
-    private final List<SubcommandGroup> groups;
-    private final List<Subcommand> subcommands;
-    private final long id, guildId, applicationId, version;
-    private final boolean defaultEnabled;
-    private final Command.Type type;
-
-    public Command(JDAImpl api, Guild guild, DataObject json)
-    {
-        this.api = api;
-        this.guild = guild;
-        this.name = json.getString("name");
-        this.description = json.getString("description", "");
-        this.type = Type.fromId(json.getInt("type", 1));
-        this.id = json.getUnsignedLong("id");
-        this.defaultEnabled = json.getBoolean("default_permission");
-        this.guildId = guild != null ? guild.getIdLong() : 0L;
-        this.applicationId = json.getUnsignedLong("application_id", api.getSelfUser().getApplicationIdLong());
-        this.options = parseOptions(json, OPTION_TEST, Option::new);
-        this.groups = parseOptions(json, GROUP_TEST, SubcommandGroup::new);
-        this.subcommands = parseOptions(json, SUBCOMMAND_TEST, Subcommand::new);
-        this.version = json.getUnsignedLong("version", id);
-    }
-
-    protected static <T> List<T> parseOptions(DataObject json, Predicate<DataObject> test, Function<DataObject, T> transform)
-    {
-        return json.optArray("options").map(arr ->
-            arr.stream(DataArray::getObject)
-               .filter(test)
-               .map(transform)
-               .collect(Collectors.toList())
-        ).orElse(Collections.emptyList());
-    }
-
     /**
      * Delete this command.
      * <br>If this is a global command it may take up to 1 hour to vanish from all clients.
@@ -105,18 +58,7 @@ public class Command implements ISnowflake
      */
     @Nonnull
     @CheckReturnValue
-    public RestAction<Void> delete()
-    {
-        if (applicationId != api.getSelfUser().getApplicationIdLong())
-            throw new IllegalStateException("Cannot delete a command from another bot!");
-        Route.CompiledRoute route;
-        String appId = getJDA().getSelfUser().getApplicationId();
-        if (guildId != 0L)
-            route = Route.Interactions.DELETE_GUILD_COMMAND.compile(appId, Long.toUnsignedString(guildId), getId());
-        else
-            route = Route.Interactions.DELETE_COMMAND.compile(appId, getId());
-        return new RestActionImpl<>(api, route);
-    }
+    RestAction<Void> delete();
 
     /**
      * Edit this command.
@@ -129,12 +71,7 @@ public class Command implements ISnowflake
      */
     @Nonnull
     @CheckReturnValue
-    public CommandEditAction editCommand()
-    {
-        if (applicationId != api.getSelfUser().getApplicationIdLong())
-            throw new IllegalStateException("Cannot edit a command from another bot!");
-        return guild == null ? new CommandEditActionImpl(api, getId()) : new CommandEditActionImpl(guild, getId());
-    }
+    CommandEditAction editCommand();
 
     /**
      * Retrieves the {@link CommandPrivilege CommandPrivileges} for this command.
@@ -155,11 +92,7 @@ public class Command implements ISnowflake
      */
     @Nonnull
     @CheckReturnValue
-    public RestAction<List<CommandPrivilege>> retrievePrivileges(@Nonnull Guild guild)
-    {
-        Checks.notNull(guild, "Guild");
-        return guild.retrieveCommandPrivilegesById(id);
-    }
+    RestAction<List<CommandPrivilege>> retrievePrivileges(@Nonnull Guild guild);
 
     /**
      * Updates the list of {@link CommandPrivilege CommandPrivileges} for this command.
@@ -184,13 +117,7 @@ public class Command implements ISnowflake
      */
     @Nonnull
     @CheckReturnValue
-    public RestAction<List<CommandPrivilege>> updatePrivileges(@Nonnull Guild guild, @Nonnull Collection<? extends CommandPrivilege> privileges)
-    {
-        if (applicationId != api.getSelfUser().getApplicationIdLong())
-            throw new IllegalStateException("Cannot update privileges for a command from another bot!");
-        Checks.notNull(guild, "Guild");
-        return guild.updateCommandPrivilegesById(id, privileges);
-    }
+    RestAction<List<CommandPrivilege>> updatePrivileges(@Nonnull Guild guild, @Nonnull Collection<? extends CommandPrivilege> privileges);
 
     /**
      * Updates the list of {@link CommandPrivilege CommandPrivileges} for this command.
@@ -215,11 +142,7 @@ public class Command implements ISnowflake
      */
     @Nonnull
     @CheckReturnValue
-    public RestAction<List<CommandPrivilege>> updatePrivileges(@Nonnull Guild guild, @Nonnull CommandPrivilege... privileges)
-    {
-        Checks.noneNull(privileges, "CommandPrivileges");
-        return updatePrivileges(guild, Arrays.asList(privileges));
-    }
+    RestAction<List<CommandPrivilege>> updatePrivileges(@Nonnull Guild guild, @Nonnull CommandPrivilege... privileges);
 
     /**
      * Returns the {@link JDA JDA} instance of this Command
@@ -227,10 +150,7 @@ public class Command implements ISnowflake
      * @return the corresponding JDA instance
      */
     @Nonnull
-    public JDA getJDA()
-    {
-        return api;
-    }
+    JDA getJDA();
 
     /**
      * The {@link Type} of command
@@ -238,10 +158,7 @@ public class Command implements ISnowflake
      * @return The command type
      */
     @Nonnull
-    public Type getType()
-    {
-        return type;
-    }
+    Type getType();
 
     /**
      * The name of this command.
@@ -249,10 +166,7 @@ public class Command implements ISnowflake
      * @return The name
      */
     @Nonnull
-    public String getName()
-    {
-        return name;
-    }
+    String getName();
 
     /**
      * The description of this command.
@@ -260,20 +174,14 @@ public class Command implements ISnowflake
      * @return The description, empty for context menu commands
      */
     @Nonnull
-    public String getDescription()
-    {
-        return description;
-    }
+    String getDescription();
 
     /**
      * Whether this command is enabled for everyone by default.
      *
      * @return True, if everyone can use this command by default.
      */
-    public boolean isDefaultEnabled()
-    {
-        return defaultEnabled;
-    }
+    boolean isDefaultEnabled();
 
     /**
      * The {@link Option Options} of this command.
@@ -281,10 +189,7 @@ public class Command implements ISnowflake
      * @return Immutable list of command options
      */
     @Nonnull
-    public List<Option> getOptions()
-    {
-        return options;
-    }
+    List<Option> getOptions();
 
     /**
      * The {@link Subcommand Subcommands} of this command.
@@ -292,10 +197,7 @@ public class Command implements ISnowflake
      * @return Immutable list of subcommands
      */
     @Nonnull
-    public List<Subcommand> getSubcommands()
-    {
-        return subcommands;
-    }
+    List<Subcommand> getSubcommands();
 
     /**
      * The {@link SubcommandGroup SubcommandGroups} of this command.
@@ -303,20 +205,14 @@ public class Command implements ISnowflake
      * @return Immutable list of subcommand groups
      */
     @Nonnull
-    public List<SubcommandGroup> getSubcommandGroups()
-    {
-        return groups;
-    }
+    List<SubcommandGroup> getSubcommandGroups();
 
     /**
      * The id of the application this command belongs to.
      *
      * @return The application id
      */
-    public long getApplicationIdLong()
-    {
-        return applicationId;
-    }
+    long getApplicationIdLong();
 
     /**
      * The id of the application this command belongs to.
@@ -324,9 +220,9 @@ public class Command implements ISnowflake
      * @return The application id
      */
     @Nonnull
-    public String getApplicationId()
+    default String getApplicationId()
     {
-        return Long.toUnsignedString(applicationId);
+        return Long.toUnsignedString(getApplicationIdLong());
     }
 
     /**
@@ -338,10 +234,7 @@ public class Command implements ISnowflake
      *
      * @see #getTimeModified()
      */
-    public long getVersion()
-    {
-        return version;
-    }
+    long getVersion();
 
     /**
      * The time this command was updated last.
@@ -351,40 +244,12 @@ public class Command implements ISnowflake
      * @see #getVersion()
      */
     @Nonnull
-    public OffsetDateTime getTimeModified()
+    default OffsetDateTime getTimeModified()
     {
         return TimeUtil.getTimeCreated(getVersion());
     }
 
-    @Override
-    public long getIdLong()
-    {
-        return id;
-    }
-
-    @Override
-    public String toString()
-    {
-        return "C:" + getName() + "(" + getId() + ")";
-    }
-
-    @Override
-    public boolean equals(Object obj)
-    {
-        if (obj == this)
-            return true;
-        if (!(obj instanceof Command))
-            return false;
-        return id == ((Command) obj).id;
-    }
-
-    @Override
-    public int hashCode()
-    {
-        return Long.hashCode(id);
-    }
-
-    public enum Type
+    enum Type
     {
         UNKNOWN(-1),
         SLASH(1),
@@ -421,7 +286,7 @@ public class Command implements ISnowflake
      * @see net.dv8tion.jda.api.interactions.commands.build.OptionData#addChoices(Command.Choice...)
      * @see net.dv8tion.jda.api.interactions.commands.build.OptionData#addChoices(Collection)
      */
-    public static class Choice
+    class Choice
     {
         private final String name;
         private long intValue = 0;
@@ -598,7 +463,7 @@ public class Command implements ISnowflake
     /**
      * An Option for a command.
      */
-    public static class Option
+    class Option
     {
         private final String name, description;
         private final int type;
@@ -771,7 +636,7 @@ public class Command implements ISnowflake
     /**
      * An Subcommand for a command.
      */
-    public static class Subcommand
+    class Subcommand
     {
         private final String name, description;
         private final List<Option> options;
@@ -780,7 +645,7 @@ public class Command implements ISnowflake
         {
             this.name = json.getString("name");
             this.description = json.getString("description");
-            this.options = parseOptions(json, OPTION_TEST, Option::new);
+            this.options = CommandImpl.parseOptions(json, CommandImpl.OPTION_TEST, Option::new);
         }
 
         /**
@@ -843,7 +708,7 @@ public class Command implements ISnowflake
     /**
      * An Subcommand Group for a command.
      */
-    public static class SubcommandGroup
+    class SubcommandGroup
     {
         private final String name, description;
         private final List<Subcommand> subcommands;
@@ -852,7 +717,7 @@ public class Command implements ISnowflake
         {
             this.name = json.getString("name");
             this.description = json.getString("description");
-            this.subcommands = parseOptions(json, SUBCOMMAND_TEST, Subcommand::new);
+            this.subcommands = CommandImpl.parseOptions(json, CommandImpl.SUBCOMMAND_TEST, Subcommand::new);
         }
 
         /**
